+++ ui/src/interfaces/users.interfaces.ts
@@ -1,0 +1,10 @@
+export interface IUser {
+    name: string
+    identification: string
+    birthday: string
+}
+
+export interface ILoginPayload {
+    identification: string
+    password: string
+}
